deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
plotly
pillow
scikit-learn==1.5.1